V1.0
Operates the same as the python code within 10e-3, fixed shifting issues, added option for LPF cutoff frequency.

+++ Window.h
@@ -15,16 +15,16 @@
 class window {
     const float PI;
     float cutoff;
-    int SamplingRate;
+    float SamplingRate;
     float Nyq;
     int N;
     float beta;
     std::vector<float> FIR;
     float StopGain;
-    float TranWidth;
+    double TranWidth;
     public:
         window();
-        window(float co, float sr, float sg, float tw);
+        window(float co, float sr, float sg, double tw);
         void kaiserord (float,float);
         float kaiser_atten (int,int);
         void kaiser_beta (float);
